refactor and close button
